# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PureGold Trading API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product catalog operations
  - name: cart
    description: Shopping cart operations
  - name: prices
    description: Live metal spot prices
  - name: orders
    description: Order operations
  - name: blog
    description: Blog and content operations
  - name: categories
    description: Product category operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products with filters
      parameters:
        - name: metal
          in: query
          schema:
            type: string
            enum: [gold, silver, platinum, palladium, copper]
        - name: category
          in: query
          schema:
            type: string
        - name: mint
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: inStock
          in: query
          schema:
            type: boolean
        - name: isIRAEligible
          in: query
          schema:
            type: boolean
        - name: isOnSale
          in: query
          schema:
            type: boolean
        - name: isNew
          in: query
          schema:
            type: boolean
        - name: isFeatured
          in: query
          schema:
            type: boolean
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [price_asc, price_desc, premium_asc, newest, best_selling, highest_rated]
        - name: limit
          in: query
          schema:
            type: integer
            default: 40
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"

  /products/featured:
    get:
      operationId: getFeaturedProducts
      tags: [products]
      summary: Get featured products for homepage
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/best-sellers:
    get:
      operationId: getBestSellers
      tags: [products]
      summary: Get best-selling products
      parameters:
        - name: metal
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: Best selling products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/on-sale:
    get:
      operationId: getOnSaleProducts
      tags: [products]
      summary: Get products on sale
      responses:
        "200":
          description: Sale products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/new-arrivals:
    get:
      operationId: getNewArrivals
      tags: [products]
      summary: Get newly arrived products
      responses:
        "200":
          description: New arrival products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/{slug}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          description: Product not found

  /products/{slug}/related:
    get:
      operationId: getRelatedProducts
      tags: [products]
      summary: Get related products
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /prices/spot:
    get:
      operationId: getSpotPrices
      tags: [prices]
      summary: Get current spot prices for all metals
      responses:
        "200":
          description: Current spot prices
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpotPricesResponse"

  /prices/history/{metal}:
    get:
      operationId: getPriceHistory
      tags: [prices]
      summary: Get historical price data for a metal
      parameters:
        - name: metal
          in: path
          required: true
          schema:
            type: string
            enum: [gold, silver, platinum, palladium, copper, bitcoin]
        - name: timeframe
          in: query
          schema:
            type: string
            enum: [1d, 1w, 1m, 3m, 1y, 5y, all]
            default: 1m
      responses:
        "200":
          description: Price history data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceHistoryResponse"

  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get current cart
      responses:
        "200":
          description: Cart contents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /cart/{productId}:
    put:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCartItemBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: Get all product categories with product counts
      responses:
        "200":
          description: Category list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /blog/posts:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List blog posts
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Blog posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"

  /blog/posts/{slug}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get blog post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"

  /stats/summary:
    get:
      operationId: getSiteSummary
      tags: [products]
      summary: Get site-wide summary stats
      responses:
        "200":
          description: Summary statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Product:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        year:
          type: integer
        mint:
          type: string
        metalType:
          type: string
          enum: [gold, silver, platinum, palladium, copper]
        purity:
          type: string
        weight:
          type: number
        weightUnit:
          type: string
          enum: [oz, g, kg, lb]
        spotPrice:
          type: number
        premiumPct:
          type: number
        premiumAmt:
          type: number
        price:
          type: number
        msrp:
          type: number
        inStock:
          type: boolean
        stockQty:
          type: integer
        isOnSale:
          type: boolean
        saleLabel:
          type: string
        isFeatured:
          type: boolean
        isNew:
          type: boolean
        isIRAEligible:
          type: boolean
        isNumismatic:
          type: boolean
        isPreOrder:
          type: boolean
        images:
          type: array
          items:
            type: string
        shortDescription:
          type: string
        category:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
      required:
        - id
        - slug
        - name
        - metalType
        - weight
        - weightUnit
        - price
        - inStock
        - images
        - category

    ProductDetail:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          properties:
            description:
              type: string
            specifications:
              type: object
              properties:
                diameter:
                  type: string
                thickness:
                  type: string
                edgeType:
                  type: string
                designer:
                  type: string
                obverse:
                  type: string
                reverse:
                  type: string
            grades:
              type: array
              items:
                type: string
            mintCertification:
              type: string
            assayCertificate:
              type: boolean
            shippingClass:
              type: string
            relatedProductIds:
              type: array
              items:
                type: string

    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
      required:
        - products
        - total
        - offset
        - limit

    SpotPrice:
      type: object
      properties:
        metal:
          type: string
        symbol:
          type: string
        price:
          type: number
        change:
          type: number
        changePercent:
          type: number
        direction:
          type: string
          enum: [up, down, flat]
        unit:
          type: string
        updatedAt:
          type: string
      required:
        - metal
        - symbol
        - price
        - change
        - changePercent
        - direction
        - unit
        - updatedAt

    SpotPricesResponse:
      type: object
      properties:
        prices:
          type: array
          items:
            $ref: "#/components/schemas/SpotPrice"
        timestamp:
          type: string
      required:
        - prices
        - timestamp

    PricePoint:
      type: object
      properties:
        date:
          type: string
        price:
          type: number
      required:
        - date
        - price

    PriceHistoryResponse:
      type: object
      properties:
        metal:
          type: string
        timeframe:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/PricePoint"
        currentPrice:
          type: number
        minPrice:
          type: number
        maxPrice:
          type: number
      required:
        - metal
        - timeframe
        - data
        - currentPrice
        - minPrice
        - maxPrice

    CartItem:
      type: object
      properties:
        productId:
          type: string
        slug:
          type: string
        name:
          type: string
        price:
          type: number
        quantity:
          type: integer
        image:
          type: string
        metalType:
          type: string
        weight:
          type: number
        weightUnit:
          type: string
      required:
        - productId
        - name
        - price
        - quantity

    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        totalItems:
          type: integer
      required:
        - items
        - subtotal
        - totalItems

    AddToCartBody:
      type: object
      properties:
        productId:
          type: string
        quantity:
          type: integer
          default: 1
      required:
        - productId

    UpdateCartItemBody:
      type: object
      properties:
        quantity:
          type: integer
      required:
        - quantity

    OrderItem:
      type: object
      properties:
        productId:
          type: string
        name:
          type: string
        price:
          type: number
        quantity:
          type: integer
        image:
          type: string
      required:
        - productId
        - name
        - price
        - quantity

    ShippingAddress:
      type: object
      properties:
        name:
          type: string
        address1:
          type: string
        address2:
          type: string
        city:
          type: string
        state:
          type: string
        zip:
          type: string
        country:
          type: string
      required:
        - name
        - address1
        - city
        - state
        - zip

    Order:
      type: object
      properties:
        id:
          type: string
        orderNumber:
          type: string
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        shipping:
          type: number
        tax:
          type: number
        total:
          type: number
        shippingAddress:
          $ref: "#/components/schemas/ShippingAddress"
        createdAt:
          type: string
        estimatedDelivery:
          type: string
      required:
        - id
        - orderNumber
        - status
        - items
        - subtotal
        - total
        - createdAt

    CreateOrderBody:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: string
              quantity:
                type: integer
            required:
              - productId
              - quantity
        shippingAddress:
          $ref: "#/components/schemas/ShippingAddress"
        paymentMethod:
          type: string
          enum: [check, wire, credit_card, bitcoin]
      required:
        - items
        - shippingAddress
        - paymentMethod

    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        metal:
          type: string
        productCount:
          type: integer
        description:
          type: string
      required:
        - id
        - name
        - slug
        - productCount

    BlogPost:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        title:
          type: string
        excerpt:
          type: string
        content:
          type: string
        author:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        publishedAt:
          type: string
        readTime:
          type: integer
        image:
          type: string
      required:
        - id
        - slug
        - title
        - excerpt
        - publishedAt

    SiteSummary:
      type: object
      properties:
        totalProductsSold:
          type: string
        totalRevenue:
          type: string
        totalCustomers:
          type: string
        totalReviews:
          type: string
        averageRating:
          type: number
        goldPrice:
          type: number
        silverPrice:
          type: number
        productsInStock:
          type: integer
        onSaleCount:
          type: integer
      required:
        - totalProductsSold
        - totalRevenue
        - totalCustomers
        - totalReviews
        - averageRating
        - goldPrice
        - silverPrice
        - productsInStock
        - onSaleCount