# lib/api-zod/src/generated/types/blogPost.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PureGold Trading API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BlogPost {
  id: string;
  slug: string;
  title: string;
  excerpt: string;
  content?: string;
  author?: string;
  category?: string;
  tags?: string[];
  publishedAt: string;
  readTime?: number;
  image?: string;
}