# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PureGold Trading API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CartItem {
  productId: string;
  slug?: string;
  name: string;
  price: number;
  quantity: number;
  image?: string;
  metalType?: string;
  weight?: number;
  weightUnit?: string;
}