# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PureGold Trading API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderBodyItemsItem } from "./createOrderBodyItemsItem";
import type { CreateOrderBodyPaymentMethod } from "./createOrderBodyPaymentMethod";
import type { ShippingAddress } from "./shippingAddress";

export interface CreateOrderBody {
  items: CreateOrderBodyItemsItem[];
  shippingAddress: ShippingAddress;
  paymentMethod: CreateOrderBodyPaymentMethod;
}